+++ lib/core/interfaces/interfaces.ts
@@ -1,0 +1,6 @@
+import { Point } from '../shapes/point';
+
+export interface ILineLineCollisionObject {
+  isColliding: boolean;
+  pointOfIntersection?: Point;
+}