+++ CHANGELOG.md
@@ -25,7 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Point-Line collision
 - Add Circle-Line collision
 - Add Line-Line collision
-- Add following shapes: Circle, Point, Rectangle, Line
+- Add shapes: Circle, Point, Rectangle, Line
+- Add interfaces: ILineLineCollisionObject
 - Add src folder with main.ts
 - Add .gitignore, .npmignore, .prettierrc, CHANGELOG.md, CODE_OF_CONDUCT.md
 - Add LICENCE, package.json, README.md and tsconfig.json